+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Adds preliminary support for the Matrix v1.1 client/v3 API
+- Adds some support for knocking rooms
 
 ## 2.4.0 - 2021-07-19